Implement a `@_resilientFinal` attribute.
The main distinction between this attribute and `final` is that
this attribute can be resiliently removed in the future when a method
wants to allow overriding, whereas `final` locks down the semantics so
that it can never be removed without breaking ABI.  Since this attribute
is inconsistent with `open`, the only reason to adopt this attribute is
to limit the code within your own module.  However, that's exactly what
we want for the unownedExecutor property of actors: when the property
is implicitly defined (e.g. for default actors), and the actor isn't
frozen, we want to reserve ABI flexibility for the actor implementation
to be changed in the future, while still generally not allowing the
actor implementation to be altered by subclasses (even within the
module) because doing so would significantly pessimize the
implementation.

The current implementation does not emit method descriptor symbols
for the method and so does not support super dispatch.  My thinking
here is that super dispatch is an extremely rarely-used feature in
the absence of outside of overrides, which are illegal; my inclination
is to also ban the use of super dispatch with `@_resilientFinal`
instead of adding overhead just to support this corner case.  But
I haven't done that yet; if you do a resilient super dispatch, you'll
just get a linker failure.